src/third_party/blink/web_tests/external/wpt/mathml/presentation-markup/mrow/mrow-preferred-width-expected.txt

This is a testharness.js-based test.
FAIL Preferred width of mrow with mspace children assert_approx_equals: unknown expected 15 +/- 1 but got 30
FAIL Preferred width of mrow with mn and mo children assert_true: expected true got false
Harness: the test ran to completion.
